==> PRS_PheWAS/2_PRS_PheWAS_phenotype.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(data.table))


master_phe_info_f <- '@@@@@@@@@@@@@@@@@@@'


master_phe_info_df <- fread(master_phe_info_f) %>%
rename('GBE_ID' = '#GBE_ID')

# This file is a copy from 
# UK Biobank Category 1307
# Infectious Disease Antigens - Infectious Diseases - Blood assays - Assay results - Biological samples
# https://biobank.ctsu.ox.ac.uk/crystal/label.cgi?id=1307

list_of_fields <- fread('UKB_Category_1307.tsv') %>%
rename('Field_ID' = '#Field_ID')


library(googlesheets)
gs_auth(token = "@@@@@@@@@@@@@@@@@@@")

file <- '@@@@@@@@@@@@@@@@@@@'
GBE_names_df <- file %>% gs_url() %>% gs_read(ws = 'GBE_names')


PheWAS_traits <- master_phe_info_df %>% 
filter(FIELD %in% list_of_fields$Field_ID) %>%
left_join(
    GBE_names_df %>%
    rename('phenotype' = 'GBE_short_name', 'phenotype_full' = 'GBE_NAME') %>%
    select(GBE_ID, phenotype, phenotype_full),
    by='GBE_ID'
) %>%
rename('N_WB' = 'N_GBE', 'UKB_Field' = 'FIELD') %>%
select(GBE_ID, phenotype, phenotype_full, UKB_Field, N, N_WB, N_AFR, N_SAS, N_EAS)


setdiff(list_of_fields$Field_ID, PheWAS_traits$UKB_Field)

PheWAS_traits %>%
rename('#GBE_ID' = 'GBE_ID') %>%
fwrite('PRS_PheWAS.phenotypes.tsv', sep='\t')
