==> PRS_PheWAS/5_PRS_PheWAS_summary.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(data.table))


#inputs
phe_info_f <- '../snpnet/pheno.info.tsv'
PheWAS_traits_f <- 'PRS_PheWAS.phenotypes.tsv'
PRS_PheWAS_res_f <- 'data/phewas.white_british.full.tsv'

# output
PRS_PheWAS_summary_f <- 'phewas.white_british.summary.tsv'


# list of target traits
PheWAS_traits <- fread(PheWAS_traits_f, sep='\t') %>%
rename('GBE_ID' = '#GBE_ID')


# read phenotype meta data file
phe_info_df <- fread(phe_info_f) %>%
rename('GBE_ID' = '#GBE_ID')


traits_for_write_up <- c(
    'INI30190',
    'INI30150',
    'INI30210',
    'INI30120',
    'INI30180',
    'INI30140',
    'INI30200',
    'INI30000'
)

phe_info_df %>%
filter(GBE_ID %in% traits_for_write_up)

PRS_PheWAS_res_df <- fread(
    PRS_PheWAS_res_f,
    colClasses=c('BETA_str'='character', 'P_str'='character')
) %>%
rename('phenotype' = '#phenotype') %>%
mutate(
    BETA_str = sprintf('%.3f (%.3f,%.3f)', BETA, l_err, u_err),
) %>%
select(-l_err, -u_err, -std_err, -BETA) %>%
rename('BETA'='BETA_str') %>%
filter(PRS_col %in% (traits_for_write_up %>% lapply(function(x){paste0(x, '_PRS')})))


PRS_PheWAS_res_df$fdr <- p.adjust(PRS_PheWAS_res_df$P, method = 'BH')


PRS_PheWAS_res_df$fdr %>% min()

PRS_PheWAS_res_df %>% 
select(phenotype, PRS_col) %>%
unique() %>% nrow()

We have 352 unique combination of PRS phenotypes and the target (Infectious Diseases) phenotypes.


PRS_PheWAS_summary_df <- PRS_PheWAS_res_df %>% 
arrange(P) %>%
select(-P) %>%
rename('P' = 'P_str') %>%
rename('GBE_ID' = 'phenotype')%>%
left_join(
    PheWAS_traits %>% select(GBE_ID, phenotype, N_WB),
    by='GBE_ID'
)%>%
left_join(
    phe_info_df %>%
    mutate(
        PRS_col = paste0(GBE_ID, '_PRS')
    ) %>%
    select(PRS_col, pheno_plot) %>%
    rename('PRS_pheno'='pheno_plot'),
    by='PRS_col'
) %>%
rename('PRS'='PRS_col') %>%
select(GBE_ID, PRS, phenotype, PRS_pheno, BETA, P, fdr)

PRS_PheWAS_summary_df %>%
rename('#GBE_ID' = 'GBE_ID') %>%
fwrite(PRS_PheWAS_summary_f, sep='\t', na = "NA", quote=F)
